fix ural vj status parsing
The author filter is no longer supported by the status page, so we have get runid in the response header and use that to locate the right row. Currently it will fetch only 10 runs initailly, if our runid cannot be found then the number of runs fetched will increase in 10 until timeout.

I also update the regex and language table, the latter fix the java problem